features: See Commits (#140)
* Add Reset Opcodes

* Add base for STOP Opcode

* Return Result from componnents

* Add Get Set for PPU Registers in memory

* Correct timings and add get and set in PPU

* Basic PPU Registers ui in Debugger

* Add Register Abstraction in UI

* Correct Processing unit run in soc

* Add Constants

+++ crates/shared/src/error.rs
@@ -4,8 +4,8 @@ use std::fmt;
 pub enum Error {
     InvalidPC(u16),
     IllegalSet(usize, u8),
-    InvalidGet(u16),
-    InvalidSet(u16, u8),
+    InvalidGet(usize),
+    InvalidSet(usize, u8),
     InvalidInterupt(u32),
     DisabledInterrupts,
     RamLock(usize),
